+++ hexital/analysis/utils.py
@@ -8,10 +8,7 @@ def realbody_avg(candles: List[Candle], length: int, index: Optional[int] = None
         index = len(candles) - 1
 
     return (
-        sum(
-            candles[i].realbody()
-            for i in range(0 if index - length < 0 else index - length, index)
-        )
+        sum(candles[i].realbody for i in range(0 if index - length < 0 else index - length, index))
         / length
     )
 
@@ -22,7 +19,7 @@ def shadow_upper_avg(candles: List[Candle], length: int, index: Optional[int] = 
 
     return (
         sum(
-            candles[i].shadow_upper()
+            candles[i].shadow_upper
             for i in range(0 if index - length < 0 else index - length, index)
         )
         / length
@@ -35,7 +32,7 @@ def shadow_lower_avg(candles: List[Candle], length: int, index: Optional[int] = 
 
     return (
         sum(
-            candles[i].shadow_lower()
+            candles[i].shadow_lower
             for i in range(0 if index - length < 0 else index - length, index)
         )
         / length
@@ -47,9 +44,6 @@ def high_low_avg(candles: List[Candle], length: int, index: Optional[int] = None
         index = len(candles) - 1
 
     return (
-        sum(
-            candles[i].high_low()
-            for i in range(0 if index - length < 0 else index - length, index)
-        )
+        sum(candles[i].high_low for i in range(0 if index - length < 0 else index - length, index))
         / length
     )