+++ CHANGELOG.md
@@ -7,15 +7,22 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 ## 1.1.0 - TBD
 
 - Removed Movement/Patterns from Hexital/Indicator
+- Changed several methods to properties in Candle (positive, realbody, etc)
+- Updated Candle to be more flexible with values
+- Renamed candles_all to get_candles
+- Removed get_indicator from Hexital as indicator already exists
+- Updated prev_exists to accept optional Indicator to be more useful
 - Corrected correlation_coefficient
-- Altered Indicator name reading, to avoid issue's such as "TR" in "ATR" 
+- Altered Indicator name reading, to avoid issue's such as "TR" in "ATR"
 - Added Counter Indicator
 - Added HMA Indicator
 
 ## 1.0.1 - 2024-04-08
+
 - Fixed #12 Inaccurate verify_indicators method in Hexital
 
 ## 1.0.0 - 2024-02-11
+
 - Moving Into BETA
 - Added CandleManager
   - CandleManger replaces List[Candle] to manage Candles and controls CandlestickTypes, timeframes and lifespan
@@ -43,6 +50,7 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 - Renamed utils/candlesticks to utils/candles
 
 ## 0.4.0 - 2024-01-22
+
 - Cleaned up code and some potential Bugs ruff/pyright
 - Added movement Above/Below and updated others to use it
 - Updated collapse candle 'fill' to show essentially doji candle rather than copy prev
@@ -61,6 +69,7 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 - Fixed purge not correctly purge sub and managed indicators
 
 ## 0.3.1 - 2023-10-09
+
 - Added candles_timerange to auto remove older than N candles
 - Added 'Settings' propety, to output Indicator in a dict format, that can be fed into back into Hexital
 - Updated Hexital to better take Pattern's as a dict input
@@ -68,14 +77,15 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 - Fixed bug in Movement on to few candles
 - Fixed bug where no timeframe indicator wasnt creating new copy of candles
 - Fixed bug where collapsed timeframe candles will use first calculated indicator value
-  - Meaning 1 minute candle that only had first 10 seconds,  will never re-calculate for rest of the minute
+  - Meaning 1 minute candle that only had first 10 seconds, will never re-calculate for rest of the minute
 
 ## 0.3.0 - 2023-09-27
+
 - Added Patterns:
   - Doji Candle
 - Added append method to Indicator just like Hexital
 - Fixed bug where Hexital would alter indicator list
-- Fixed bug in _find_calc_index with no candles
+- Fixed bug in \_find_calc_index with no candles
 - Added Pattern Indicator, skeleton to run Any Patterns as a Indicator
   - E.G On all Candles automatically
 - Added support to generate Indicators on multiple timeframes at once
@@ -83,10 +93,11 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
   - E.G 1m candles can be used to generate 10m indicators simultaneously with 1m indicators
 
 ## 0.2.0 - 2023-09-05
+
 - Feature: Added timestamp (datetime) to Candle dataclass
 - Feature: Can convert Candle from list and dict
 - Feature: Can set timestamp(datetime) in lists/dict Candle conversion
-- Feature: Added _validate_fields method to Indicators
+- Feature: Added \_validate_fields method to Indicators
 - Added custom exceptions
 - More thorough unit testing
 - Updated Indicators accuracy to Truth source
@@ -105,9 +116,11 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
   - VWAP
   - VMA
   - WMA
-  
+
 ## 0.1.1 - 2023-08-28
+
 - Nada
 
 ## 0.1.0 - 2023-08-27
+
 - Alpha release `hexital`